The DatapathController creates 3 tunnel ports
This patch is the first patch in the vxlan overlay tunnelling feature.
It keeps the status quo and only changes how midolman set up the
datapath at startup. With this patch, host2host tunnelling is still
using the gre protocol.

Now, 2 vxlan ports are always created at startup. One is for handling
tunnelled traffic coming from peer vteps. The vxlan udp port is set by
default to the standard value, but can be changed in the configuration.
The second vxlan port is for handling overlay tunnelled traffic from
another midolman host. The vxlan udp port is also set in the
configuration. This requires that midolman talks to an ovs datapath
version 1.10 or later.

Ref MN-1867
Signed-off-by: hugobenichi <hugo@midokura.com>
Change-Id: I45b29b495dbd2402e061d6ad1cdd85742d0691ef