VxGW: make mmm test work
This requires a few changes in the MMM setup, mostly reverting previous
additions for this test. For now, the test assumes that Midolman 1
inside MMM has direct L2 connectivity to the remote VM where the
emulator runs. That is, vxlan traffic will flow through MMM's br0, and
you must configure the host so that this traffic gets out of the VM and
reaches the remote VM with the emulator. In the MidoCloud this required
adding an additional network 10.0.0.0/24 and putting both the emulator
and the MMM box. Then on the MMM box, assuming the interface plugged to
that network is eth1:

- Remove the 10.0.0.0/24 ip from the interface
- Add eth1 to br0
- Remove any routes sending 10.0.0.0/24 traffic to eth1 (these interfere
  with MMM)
- ip route add 10.0.0.101 dev br0 (assuming the vtep is at 10.0.0.101)

This also ammends the tests, mostly to simplify, and removes old cases
used for troublehooting. Also incorporates the extra vtep tunnel zone
which is required to be able to talk to configure the vxlan tunnel.

Change-Id: Ice4caaab38144b197db51cf900a32aa345702c09
Fixes: MN-2634
Signed-off-by: Galo Navarro <galo@midokura.com>